requirements.txt
# Core dependencies  
faster-whisper==1.0.3          # Speech recognition (optimized)
ollama==0.4.7                  # LLM interface
openai==1.63.2                 # OpenAI API client
requests==2.32.3               # HTTP library
rich==13.9.4                   # Terminal UI
sounddevice==0.5.3             # Audio capture
numpy==2.3.5                  # Audio processing

# Audio processing
# pyaudio==0.2.14                # Audio I/O
webrtcvad==2.0.10              # Voice activity detection

# Configuration
pyyaml==6.0.3                  # Config file parsing
python-dotenv==1.2.1           # Environment variables
pydantic==2.12.5                # Data validation

# Text-to-Speech
pyttsx3==2.99                  # Fallback TTS engine
edge-tts==7.2.7                # Microsoft Edge TTS

# Web interface
python-socketio[asyncio]==5.15.1 # Socket.IO
fastapi==0.125.0               # Modern async web framework
uvicorn[standard]==0.38.0      # ASGI server
jinja2==3.1.6                 # Template engine
aiofiles==25.1.0              # Async file I/O
httpx==0.28.1                  # Async HTTP client

# Utilities
keyboard==0.13.5               # Keyboard input detection
Pillow==11.1.0                 # Image processing

# ChatLink dependencies (Unified Backend)
websocket-client               # Signal WebSocket client
google-generativeai            # Google Gemini API

# Testing
pytest==9.0.2                  # Testing framework
pytest-cov==7.0.0              # Coverage reporting
